feat(dasclaw_mcp): migrate session + transport from ironclaw (F3.2 phase 3 PR 6-α)
把 ironclaw 中两个零业务依赖的 MCP 模块搬到 dasclaw_mcp，作为 F3.2 phase 3
的第一个切片（用户反馈拆 6 个 PR 太散后合并的 3-PR 计划之 PR 1）：

- session: McpSession / McpSessionManager（管理 Mcp-Session-Id header 和
  会话活跃度），零 ironclaw 依赖，按原样移过来。
- transport: McpTransport trait + write_jsonrpc_line / spawn_jsonrpc_reader /
  stream_transport_send（JSON-RPC 帧封装和待响应分发），导入路径调整为
  crate::protocol::{McpRequest,McpResponse} 和 dasclaw_tool::ToolError。
  stream_transport_send 可见性从 pub(crate) 提升到 pub，因为 stdio /
  unix 两个具体 transport 仍在 ironclaw（PR 6-β 才搬），需要跨 crate 调用。

ironclaw 侧两个旧文件替换为 1-行 pub use shim，保持
crate::tools::mcp::{session,transport}::* 这一路径可继续解析，零 import 改动。

依赖增量（dasclaw_mcp/Cargo.toml）：
- async-trait = "0.1"
- tokio 增加 io-util / rt / time features

已检查 X 是否已有，结论：dasclaw_tool::ToolError 已是 host-agnostic、
dasclaw_runtime 已被 dasclaw_mcp 依赖，无需新建抽象层。

验证：
- cargo check -p dasclaw_mcp --tests: 0 错 0 警
- cargo check -p dasclaw --tests: 0 错 0 警（5m35s，主要是 sccache miss 后重编）
- cargo nextest run -p dasclaw_mcp: 139/139 通过（含 22 个 session + 4 个 transport）
- cargo clippy --no-deps -p dasclaw_mcp --all-targets -- -D warnings: 0 警
- cargo fmt --all: clean
- scripts/check_no_panics.py --base origin/xClaw: OK

Refs #661
