+++ crates/dasclaw_mcp/Cargo.toml
@@ -20,8 +20,11 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 thiserror = "1"
 # Async filesystem I/O for `load_mcp_servers_from` / `save_mcp_servers_to`,
-# plus async mutex for the per-(server,user) refresh lock in `auth`.
-tokio = { version = "1", features = ["fs", "sync"] }
+# async mutex for the per-(server,user) refresh lock in `auth`, plus
+# `io-util` for `AsyncBufReadExt::lines` / `AsyncWriteExt::write_all` used
+# by `transport`, `rt` for `tokio::spawn` of the JSON-RPC reader task, and
+# `time` for the `stream_transport_send` per-request timeout.
+tokio = { version = "1", features = ["fs", "sync", "io-util", "rt", "time"] }
 # URL parsing for the localhost / loopback detection in `config::is_localhost_url`.
 url = "2"
 # OAuth discovery / token endpoint HTTP client + RFC 9110 header validation.
@@ -34,6 +37,8 @@ chrono = { version = "0.4", features = ["serde"] }
 rand = "0.8"
 sha2 = "0.10"
 urlencoding = "2"
+# `transport::McpTransport` is an async trait (added in F3.2 phase 3 PR 6-α, #661).
+async-trait = "0.1"
 
 [dev-dependencies]
 tempfile = "3"
